test(ui): fix refresh proof lint and share palette helpers
